Updated the main page recipe card tiles and search
