+++ views/partials/nav.handlebars
@@ -22,10 +22,10 @@
       <form class="form-inline my-2 my-lg-0">
         <div class="container">
           <div class="row">
-            <input class="form-control mr-sm-2" type="search" placeholder="Search">
+            <input class="form-control mr-sm-2" type="search" placeholder="Search" id="searchTerm">
             <li class="nav-item dropdown ml-2">
               <div class="btn-group">
-                <button class="btn btn-outline-success my-2 my-sm-0" type="submit">Search</button>
+                <button class="btn btn-outline-success my-2 my-sm-0" type="submit" id="recipeSearch">Search</button>
                 <button class="btn btn-outline-success dropdown-toggle" type="button" id="navbarDropdownMenuLink"
                   data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
                   <span class="sr-only">Menu</span>
@@ -55,3 +55,5 @@
     </ul>
   </div>
 </nav>
+
+<script type="text/javascript" src="/js/search.js" charset="utf-8"></script>
