+++ views/search.handlebars
@@ -23,13 +23,8 @@
                                 <input type="proteinType" class="form-control" id="proteinType" placeholder="e.g. Chicken, fish, egg">
                             </div>
                             <div class="form-group">
-<<<<<<< HEAD
-                                <label for="veggieType">Enter Preferred Protein Type</label>
-                                <input type="veggieType" class="form-control" id="veggieType" placeholder="Green Beans">
-=======
                                 <label for="veggieType">Enter Preferred Veggie Type</label>
                                 <input type="veggieType" class="form-control" id="veggieType" placeholder="e.g. Green Beans, Eggplants, Carrot">
->>>>>>> jenay1
                             </div>
                             <div class="form-group">
                                 <label for="dietaryRestrictions">Select Dietary Restrictions:</label>
@@ -78,29 +73,6 @@
                             <button type="submit" class="searchRecipe">Search</button>
                         </form>
                     </p>
-<<<<<<< HEAD
-                    <div id="resultsTable">
-                        {{#if recipes.searched}}
-                        <h3>Search Results</h3>
-                        <div class="card-columns">
-                            {{#each recipes.searched}}
-                            <div class="card mt-2">
-                                <img src="{{this.imageSrc}}" alt="recipe image unavailable" class="card-img-top img-thumbnail rounded">
-                                <div class="card-body">
-                                    <h5 class="card-title">{{this.name}}</h5>
-                                    <p class="card-text">
-                                        <a href="{{this.description}}">{{this.description}}</a>
-                                    </p>
-                                </div>
-                            </div>
-                            {{/each}}
-                        </div>
-                        {{else}}
-                        <h3>No matches found. Please refine your search!</h3>
-                    </div>
-=======
-
->>>>>>> jenay1
                 </div>
             </div>
         </div>
