+++ views/users/index.handlebars
@@ -22,7 +22,7 @@
 {{/if}}
 
 <hr>
-{{!-- {{#if recipes.recommend}} --}}
+{{#if recipes.recommend}}
   <h3>Recommended Recipes</h3>
   <div class="card-columns">
    {{#each recipes.recommend}}
@@ -39,7 +39,7 @@
   </div>
 {{!-- {{else}} --}}
   {{!-- <h3>No recipe recommendations generated yet</h3> --}}
-{{!-- {{/if}} --}}
+{{/if}}
 
 <hr>
 {{#if recipes.posted}}
